+++ internal/product/repository.go
@@ -4,7 +4,7 @@ import "context"
 
 type ProductRepository interface {
 	Create(ctx context.Context, product *Product) error
-	GetByID(ctx context.Context, id int) (*Product, error)
-	Delete(ctx context.Context, id int) error
+	GetByID(ctx context.Context, id int64) (*Product, error)
+	Delete(ctx context.Context, id int64) error
 	List(ctx context.Context) ([]*Product, error)
 }
